Automatically build dist for dependabot
