+++ .github/workflows/check-dist.yml
@@ -18,6 +18,7 @@ on:
 
 jobs:
   check-dist:
+    name: Check dist
     runs-on: ubuntu-latest
 
     steps:
@@ -51,3 +52,33 @@ jobs:
         with:
           name: dist
           path: dist/
+
+  build-dist:
+    name: Build dist (Dependabot)
+    runs-on: ubuntu-latest
+    needs: [ check-dist ]
+    permissions:
+      contents: write
+
+    if: >-
+      ${{ 
+        always() &&
+        github.event_name == 'pull_request' &&
+        github.actor == 'dependabot[bot]' &&
+        startsWith( github.head_ref, 'dependabot/npm_and_yarn/' ) &&
+        needs.check-dist.result == 'failure'
+      }}
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          ref: ${{ github.head_ref }}
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Build and Push
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          tsc helper/dependabot.ts
+          node helper/dependabot.js
